few refactorings
