+++ package.json
@@ -1,36 +1,29 @@
 {
   "name": "videohunt",
   "version": "0.1.0",
+  "description": "VideoHunt React App",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/sajjaphani/videohunt.git"
+  },
   "private": true,
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject",
+    "start:prod": "node server.js"
+  },
   "dependencies": {
-    "autoprefixer": "7.1.6",
     "axios": "^0.16.2",
-    "babel-core": "6.26.0",
-    "babel-eslint": "7.2.3",
-    "babel-jest": "20.0.3",
-    "babel-loader": "7.1.2",
-    "babel-preset-react-app": "^3.1.0",
     "babel-runtime": "6.26.0",
-    "case-sensitive-paths-webpack-plugin": "2.1.1",
     "chalk": "1.1.3",
     "compression": "^1.7.1",
-    "css-loader": "0.28.7",
-    "dns": "^0.2.2",
     "dotenv": "4.0.0",
-    "eslint": "4.10.0",
-    "eslint-config-react-app": "^2.0.1",
-    "eslint-loader": "1.9.0",
-    "eslint-plugin-flowtype": "2.39.1",
-    "eslint-plugin-import": "2.8.0",
-    "eslint-plugin-jsx-a11y": "5.1.1",
-    "eslint-plugin-react": "7.4.0",
     "express": "^4.16.2",
     "express-http-proxy": "^1.0.7",
-    "extract-text-webpack-plugin": "3.0.2",
-    "file-loader": "1.1.5",
     "fs-extra": "3.0.1",
     "history": "^4.6.1",
-    "html-webpack-plugin": "2.29.0",
     "immutable": "^3.8.1",
     "jest": "20.0.4",
     "jsonwebtoken": "^8.1.0",
@@ -49,7 +42,7 @@
     "react-router": "^4.2.0",
     "react-router-dom": "^4.2.2",
     "react-router-redux": "^5.0.0-alpha.6",
-    "react-scripts": "^1.0.16",
+    "react-scripts": "1.1.5",
     "redux": "^3.5.2",
     "redux-logger": "^2.6.1",
     "redux-promise": "^0.5.3",
@@ -57,21 +50,34 @@
     "redux-saga": "^0.15.4",
     "redux-thunk": "^2.1.0",
     "reselect": "^3.0.1",
-    "semantic-ui-react": "^0.75.1",
-    "style-loader": "0.19.0",
-    "sw-precache-webpack-plugin": "0.11.4",
-    "url-loader": "0.6.2",
-    "webpack": "3.8.1",
-    "webpack-dev-server": "2.9.3",
-    "webpack-manifest-plugin": "1.3.2",
-    "whatwg-fetch": "2.0.3"
+    "semantic-ui-react": "^0.77.2",
+    "whatwg-fetch": "2.0.4"
   },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "start:prod": "node server.js"
+  "devDependencies": {
+    "autoprefixer": "7.1.6",
+    "babel-core": "6.26.3",
+    "babel-eslint": "7.2.3",
+    "babel-jest": "20.0.3",
+    "babel-loader": "7.1.2",
+    "babel-preset-react-app": "^3.1.0",
+    "case-sensitive-paths-webpack-plugin": "2.2.0",
+    "css-loader": "0.28.11",
+    "eslint": "4.10.0",
+    "eslint-config-react-app": "^2.0.1",
+    "eslint-loader": "1.9.0",
+    "eslint-plugin-flowtype": "2.50.3",
+    "eslint-plugin-import": "2.8.0",
+    "eslint-plugin-jsx-a11y": "5.1.1",
+    "eslint-plugin-react": "7.4.0",
+    "extract-text-webpack-plugin": "3.0.2",
+    "file-loader": "1.1.11",
+    "html-webpack-plugin": "2.29.0",
+    "style-loader": "0.23.1",
+    "sw-precache-webpack-plugin": "0.11.5",
+    "url-loader": "0.6.2",
+    "webpack": "3.11.0",
+    "webpack-dev-server": "2.11.5",
+    "webpack-manifest-plugin": "1.3.2"
   },
   "proxy": {
     "/api": {
